feat: renaming and adjusting code of identity to account
