--- code-examples/1_1_identity.ts
@@ -1,10 +1,0 @@
-import * as Kilt from '@kiltprotocol/sdk-js'
-
-export function main(): Kilt.Identity {
-  const mnemonic = Kilt.Identity.generateMnemonic()
-  console.log('mnemonic:', mnemonic)
-
-  const identity = Kilt.Identity.buildFromMnemonic(mnemonic)
-  console.log('address:', identity.address)
-  return identity
-}
